requirements.txt
llama-index
llama-index-core
llama-parse
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
chromadb
streamlit
python-dotenv
pandas
nest_asyncio
llama-index-llms-groq
psycopg2-binary
sqlalchemy